+++ src/module.js
@@ -15,8 +15,18 @@ import {
   DeathSaveReminder,
   SkillReminder,
 } from "./reminders.js";
+import { showSources } from "./settings.js";
+import {
+  AbilityCheckSource,
+  AbilitySaveSource,
+  AttackSource,
+  CriticalSource,
+  DeathSaveSource,
+  SkillSource,
+} from "./sources.js";
 import { debug, log } from "./util.js";
 
+const CIRCLE_INFO = `<i class="fa-solid fa-circle-info"></i> `;
 let checkArmorStealth;
 let skipReminders;
 
@@ -74,6 +84,10 @@ Hooks.on("dnd5e.preRollAttack", (item, config) => {
 
   debug("checking for message effects on this attack roll");
   new AttackMessage(item.actor, item).addMessage(config);
+  if (showSources) {
+    debug("checking for adv/dis effects to display their source");
+    new AttackSource(item.actor, getTarget(), item).updateOptions(config);
+  }
 
   if (skipReminders) return;
   debug("checking for adv/dis effects on this attack roll");
@@ -92,6 +106,10 @@ Hooks.on("dnd5e.preRollAbilitySave", (actor, config, abilityId) => {
 
   debug("checking for message effects on this saving throw");
   new AbilitySaveMessage(actor, abilityId).addMessage(config);
+  if (showSources) {
+    debug("checking for adv/dis effects to display their source");
+    new AbilitySaveSource(actor, abilityId).updateOptions(config);
+  }
 
   if (skipReminders) return;
   debug("checking for adv/dis effects on this saving throw");
@@ -106,6 +124,10 @@ Hooks.on("dnd5e.preRollAbilityTest", (actor, config, abilityId) => {
 
   debug("checking for message effects on this ability check");
   new AbilityCheckMessage(actor, abilityId).addMessage(config);
+  if (showSources) {
+    debug("checking for adv/dis effects to display their source");
+    new AbilityCheckSource(actor, abilityId).updateOptions(config);
+  }
 
   if (skipReminders) return;
   debug("checking for adv/dis effects on this ability check");
@@ -120,6 +142,10 @@ Hooks.on("dnd5e.preRollSkill", (actor, config, skillId) => {
 
   debug("checking for message effects on this skill check");
   new SkillMessage(actor, skillId).addMessage(config);
+  if (showSources) {
+    debug("checking for adv/dis effects to display their source");
+    new SkillSource(actor, skillId, checkArmorStealth).updateOptions(config);
+  }
 
   if (skipReminders) return;
   debug("checking for adv/dis effects on this skill check");
@@ -134,6 +160,10 @@ Hooks.on("dnd5e.preRollToolCheck", (item, config) => {
 
   debug("checking for message effects on this tool check");
   new AbilityCheckMessage(item.actor, item.system.ability).addMessage(config);
+  if (showSources) {
+    debug("checking for adv/dis effects to display their source");
+    new AbilityCheckSource(item.actor, item.system.ability).updateOptions(config);
+  }
 
   if (skipReminders) return;
   debug("checking for adv/dis effects on this tool check");
@@ -148,6 +178,10 @@ Hooks.on("dnd5e.preRollDeathSave", (actor, config) => {
 
   debug("checking for message effects on this death save");
   new DeathSaveMessage(actor).addMessage(config);
+  if (showSources) {
+    debug("checking for adv/dis effects to display their source");
+    new DeathSaveSource(actor).updateOptions(config);
+  }
 
   if (skipReminders) return;
   debug("checking for adv/dis effects on this death save");
@@ -163,6 +197,10 @@ Hooks.on("dnd5e.preRollDamage", (item, config) => {
 
   debug("checking for message effects on this damage roll");
   new DamageMessage(item.actor, item).addMessage(config);
+  if (showSources) {
+    debug("checking for adv/dis effects to display their source");
+    new CriticalSource(item.actor, getTarget(), item).updateOptions(config);
+  }
 
   if (skipReminders) return;
   debug("checking for critical/normal effects on this damage roll");
@@ -204,8 +242,21 @@ Hooks.on("renderDialog", async (dialog, html, data) => {
 });
 
 async function prepareMessage(dialogOptions) {
-  const messages = dialogOptions["adv-reminder"]?.messages ?? [];
-  const rollData = dialogOptions["adv-reminder"]?.rollData ?? {};
+  const opt = dialogOptions["adv-reminder"];
+  if (!opt) return;
+
+  // merge the messages with the advantage/disadvantage from sources
+  const messages = [...(opt.messages ?? [])];
+  const addLabels = (labels, stringId) => {
+    if (labels) {
+      const sources = labels.join(", ");
+      messages.push(CIRCLE_INFO + game.i18n.format(stringId, { sources }));
+    }
+  };
+  addLabels(opt.advantageLabels, "adv-reminder.Source.Adv");
+  addLabels(opt.disadvantageLabels, "adv-reminder.Source.Dis");
+  addLabels(opt.criticalLabels, "adv-reminder.Source.Crit");
+  addLabels(opt.normalLabels, "adv-reminder.Source.Norm");
 
   if (messages.length) {
     // build message
@@ -219,7 +270,7 @@ async function prepareMessage(dialogOptions) {
       documents: true,
       links: false,
       rolls: true,
-      rollData,
+      rollData: opt.rollData ?? {},
       async: false,
     });
     debug("messages", messages, "enriched", enriched);